# swagger.yaml
openapi: 3.0.0
info:
  title: Ticketing Platform API
  description: API for managing events, tickets, and user authentication
  version: 1.0.0
  contact:
    name: SmartTicket Support
servers:
  - url: http://localhost:${PORT}/api
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        venue_id:
          type: integer
          format: int64
        organizer_id:
          type: integer
          format: int64
        event_type_id:
          type: integer
          format: int64
        total_tickets:
          type: integer
        available_tickets:
          type: integer
        price:
          type: number
          format: float
        venue:
          $ref: '#/components/schemas/Venue'
        organizer:
          $ref: '#/components/schemas/Organizer'
        eventType:
          $ref: '#/components/schemas/EventType'
        artists:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
    Venue:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        city_id:
          type: integer
          format: int64
        capacity:
          type: integer
        description:
          type: string
          nullable: true
    Organizer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        bio:
          type: string
          nullable: true
    EventType:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
    Artist:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        genre:
          type: string
        bio:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
paths:
...
  /events/{id}:
    get:
      tags:
        - Events
      summary: Get event by ID
      description: Retrieve a specific event with all details
      parameters:
        - name: id
          in: path
          required: true
          description: Event ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Events
      summary: Update event
      description: Update an existing event's information
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Event ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                date:
                  type: string
                  format: date-time
                venue_id:
                  type: integer
                organizer_id:
                  type: integer
                event_type_id:
                  type: integer
                price:
                  type: number
                  format: float
                status:
                  type: string
                  enum: [upcoming, ongoing, completed, cancelled]
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  event:
                    $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Events
      summary: Delete event
      description: Delete an event and related data
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Event ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Event deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...